api/requirement.txt
Django
python-decouple
djangorestframework
django-cors-headers
mysqlclient
